+++ gala/potential/potential/tests/test_io.py
@@ -4,6 +4,7 @@
 import astropy.units as u
 from astropy.utils.data import get_pkg_data_filename
 import numpy as np
+import pytest
 
 # Project
 from ..io import load, save
@@ -11,7 +12,9 @@ from ..core import CompositePotential
 from ..ccompositepotential import CCompositePotential
 from ..builtin import IsochronePotential, KeplerPotential
 from ..builtin.special import LM10Potential
+from ...scf import SCFPotential
 from ....units import DimensionlessUnitSystem, galactic
+from gala._cconfig import GSL_ENABLED
 
 
 def test_read_plummer():
@@ -118,5 +121,16 @@ def test_units(tmpdir):
 
     # try a simple potential
     potential = KeplerPotential(m=1E11, units=[u.kpc, u.Gyr, u.Msun, u.radian])
+    save(potential, tmp_filename)
+    p = load(tmp_filename)
+
+
+@pytest.mark.skipif(not GSL_ENABLED,
+                    reason="requires GSL to run this test")
+def test_read_write_SCF(tmpdir):
+    tmp_filename = str(tmpdir.join("potential.yml"))
+
+    # try a basic SCF potential
+    potential = SCFPotential(100, 1, np.zeros((4, 3, 2)), np.zeros((4, 3, 2)))
     save(potential, tmp_filename)
     p = load(tmp_filename)